bc_datacleaning/requirements.txt
pandas>=1.5.0
openpyxl>=3.0.0
xlrd>=2.0.1
openai>=1.0.0
anthropic>=0.7.0
pyyaml>=6.0
pathlib2>=2.3.0
typing-extensions>=4.0.0
requests>=2.28.0
transformers>=4.30.0
torch>=2.0.0
sentence-transformers>=2.2.0
numpy>=1.21.0
urllib3<2.0.0 